// src/components/AddExpenseTable/AddExpenseTable.vue.d.ts
export type NewExpenseData = {
    date?: number;
    name?: string;
    netAmount?: number;
    amount?: number;
    paidBackAmount?: number;
    category?: string;
    subCategory?: string;
};
declare const _default: import("vue").DefineComponent<{}, {}, {}, {}, {}, import("vue").ComponentOptionsMixin, import("vue").ComponentOptionsMixin, {}, string, import("vue").PublicProps, Readonly<{}> & Readonly<{}>, {}, {}, {}, {}, string, import("vue").ComponentProvideOptions, true, {}, any>;
export default _default;